Convert DNS challenge providers to the plugin provider model (This also allows hot-adding new DNS challenge providers without having to recompile Certify)
Outstanding issues:
1. `PluginManager` (part of `Shared.Core`) has a hidden hard dependency
on `Core` now when told to load DNS providers; it specifically loads two
classes from `Core`. It's not a compile-time dependency though.

2. `PluginManager` wasn't directly reachable from the part of the code
that processes DNS challenge provider brokering. The closest place it
could be reached was in `CertifyManager`, which is 7-8 calls further up
the stack. Rather than plumb all of those calls with an added parameter
that the functions otherwise don't use, I statically expose it. I see
that passing configuration objects already has the same problem and
solution.

3. `ChallengeDiagnostics` includes a one-off correction for the
`AcmeDns` provider. Previously that was a direct reference to the
provider definition ID; since I broke those direct dependencies, that's
now a plain string.

4. `DnsProviderLibcloud` is still shipped, but is now explicitly unused.
Previously it could theoretically be selected if any Python challenge
provider definitions existed, but none did and there wasn't a way to
extend this. (It can still be referenced by consumer-created plugins,
which are now possible.

5. None of the existing DNS provider DLLs are directly referenced
anymore, so they're not automatically copied as part of the build
process. They need to be manually built and copied or else the service
will crash at runtime when `PluginManager` does to load them.